+++ chat-widget/samples/typescript-sample/src/index.tsx
@@ -2,6 +2,9 @@ import { OmnichannelChatSDK } from "@microsoft/omnichannel-chat-sdk";
 import { LiveChatWidget } from "@microsoft/omnichannel-chat-widget";
 import React, { useEffect, useState } from "react";
 import ReactDOM from "react-dom";
+import * as OcChatSdkPackageinfo from "@microsoft/omnichannel-chat-sdk/package.json";
+import * as OcChatWidgetPackageInfo from "@microsoft/omnichannel-chat-widget/package.json";
+import * as OcChatComponentPackageInfo from "@microsoft/omnichannel-chat-components/package.json";
 
 const getOmnichannelChatConfig = () => {
     // add your own OC setting, hard-coded just for sample, should be replaced with a better handling
@@ -14,9 +17,15 @@ const getOmnichannelChatConfig = () => {
 };
 
 const App = () => {
+    // To avoid webpack 5 warning and soon obsolete code, rename the packageinfo variable
+    const OcSdkPkginfo = OcChatSdkPackageinfo;
+    const OcChatWidgetPkgInfo = OcChatWidgetPackageInfo;
+    const OcChatComponentPkgInfo = OcChatComponentPackageInfo;
+    
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const [liveChatWidgetProps, setLiveChatWidgetProps] = useState<any>();
     const omnichannelConfig = getOmnichannelChatConfig();
+
     useEffect(() => {
         const init = async () => {
             const chatSDK = new OmnichannelChatSDK(omnichannelConfig);
@@ -36,6 +45,11 @@ const App = () => {
                 chatConfig,
                 webChatContainerProps: {
                     disableMarkdownMessageFormatting: true, //setting the default to true for a known issue with markdown
+                },
+                telemetryConfig: { //mandatory for telemetry
+                    chatWidgetVersion: OcChatWidgetPkgInfo.version,
+                    chatComponentVersion: OcChatComponentPkgInfo.version,
+                    OCChatSDKVersion: OcSdkPkginfo.version
                 }
             };